moving endpoint and request builder into http_file
