--- src/hat_util/error.rs
@@ -1,5 +1,0 @@
-#[derive(thiserror::Error, Debug)]
-pub enum UtilError {
-    #[error("{0}")]
-    IO(#[from] std::io::Error),
-}
