+++ src/hat_util/mod.rs
@@ -1,11 +1,3 @@
-mod endpoint;
-mod error;
-mod request_builder;
 mod store;
 
-pub use {
-    endpoint::{Endpoint, EndpointError},
-    error::UtilError,
-    request_builder::RequestBuilder,
-    store::{Store, StoreComposed, StoreUnion},
-};
+pub use store::{Store, StoreComposed, StoreUnion};
